videos

+++ database/migrations/2021_05_11_172326_create_videos_table.php
@@ -17,7 +17,7 @@ class CreateVideosTable extends Migration
             $table->id();
             $table->string("video");
             $table->string("description");
-            $table->bigInteger("position");
+            $table->Integer("position");
             $table->timestamps();
             $table->softDeletes();
             $table->bigInteger("proyects_id")->unasigned();
